+++ tests/framework/lookup/test_lookup.py
@@ -1,12 +1,15 @@
 import itertools
-from typing import Union
 
 import numpy as np
 import pandas as pd
 import pytest
 
 from vivarium import InteractiveContext
-from vivarium.framework.lookup import LookupTable, validate_build_table_parameters
+from vivarium.framework.lookup import (
+    LookupTableInterface,
+    LookupTableManager,
+    validate_build_table_parameters,
+)
 from vivarium.testing_utilities import TestPopulation, build_table
 
 
@@ -123,7 +126,7 @@ def test_interpolated_tables__exact_values_at_input_points(base_config):
     simulation = InteractiveContext(components=[TestPopulation()], configuration=base_config)
     manager = simulation._tables
     years = manager._build_table(
-        years, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=None
+        years, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=()
     )
 
     for year in input_years:
@@ -186,9 +189,7 @@ def test_invalid_data_type_build_table(base_config):
     simulation = InteractiveContext(components=[TestPopulation()], configuration=base_config)
     manager = simulation._tables
     with pytest.raises(TypeError):
-        manager._build_table(
-            "break", key_columns=None, parameter_columns=None, value_columns=None
-        )
+        manager._build_table("break", key_columns=(), parameter_columns=(), value_columns=())
 
 
 def test_lookup_table_interpolated_return_types(base_config):
@@ -199,7 +200,7 @@ def test_lookup_table_interpolated_return_types(base_config):
     simulation = InteractiveContext(components=[TestPopulation()], configuration=base_config)
     manager = simulation._tables
     table = manager._build_table(
-        data, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=None
+        data, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=()
     )(simulation.get_population().index)
     # make sure a single value column is returned as a series
     assert isinstance(table, pd.Series)
@@ -207,7 +208,7 @@ def test_lookup_table_interpolated_return_types(base_config):
     # now add a second value column to make sure the result is a df
     data["value2"] = data.value
     table = manager._build_table(
-        data, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=None
+        data, key_columns=["sex"], parameter_columns=["age", "year"], value_columns=()
     )(simulation.get_population().index)
 
     assert isinstance(table, pd.DataFrame)
@@ -227,21 +228,13 @@ def test_validate_parameters_no_data(data):
         ((), (), (), "supply value_columns"),
         ((), (), [], "supply value_columns"),
         ((), (), ["a", "b"], "match the number of values"),
+        (("a", "b"), (), ["d", "e", "f"], "key_columns are not allowed"),
+        ((), ("a", "b"), ["d", "e", "f"], "parameter_columns are not allowed"),
     ],
 )
 def test_validate_parameters_error_scalar_data(key_cols, param_cols, val_cols, match):
     with pytest.raises(ValueError, match=match):
         validate_build_table_parameters([1, 2, 3], key_cols, param_cols, val_cols)
-
-
-@pytest.mark.parametrize(
-    "key_cols, param_cols, val_cols, match",
-    [(["a", "b"], ["b"], ["c"], "no overlap"), ([], ["b"], ["c"], "do not match")],
-)
-def test_validate_parameters_error_dataframe(key_cols, param_cols, val_cols, match):
-    data = pd.DataFrame({"a": [1, 2], "b_start": [0, 5], "b_end": [5, 10], "c": [100, 150]})
-    with pytest.raises(ValueError, match=match):
-        validate_build_table_parameters(data, key_cols, param_cols, val_cols)
 
 
 @pytest.mark.parametrize(
@@ -253,53 +246,59 @@ def test_validate_parameters_fail_other_data(data):
 
 
 @pytest.mark.parametrize(
-    "key_cols, param_cols, val_cols",
+    "key_cols, param_cols, val_cols, match",
     [
-        (None, None, ["a", "b", "c"]),
-        (None, ["d"], ["one", "two", "three"]),
-        (["KEY"], None, ["a", "b", "c"]),
-        (["KEY"], ["d"], ["a", "b", "c"]),
+        ([], [], ["c"], "either key_columns or parameter_columns"),
+        (["a", "b"], ["b"], ["c"], "no overlap between key.*and parameter columns"),
+        (["a"], ["b"], ["a", "c"], "no overlap between value.*and key.*columns"),
+        (["a"], ["b"], ["b", "c"], "no overlap between value.*and.*parameter columns"),
+        (["d"], ["b"], ["c"], "columns.*must all be present"),
+        (["a"], ["d"], ["c"], "columns.*must all be present"),
+        (["a"], ["b"], ["d"], "columns.*must all be present"),
     ],
 )
-def test_validate_parameters_pass_scalar_data(key_cols, param_cols, val_cols):
-    validate_build_table_parameters([1, 2, 3], key_cols, param_cols, val_cols)
+def test_validate_parameters_error_dataframe(key_cols, param_cols, val_cols, match):
+    data = pd.DataFrame({"a": [1, 2], "b_start": [0, 5], "b_end": [5, 10], "c": [100, 150]})
+    with pytest.raises(ValueError, match=match):
+        validate_build_table_parameters(data, key_cols, param_cols, val_cols)
+
+
+def test_validate_parameters_pass_scalar_data():
+    validate_build_table_parameters([1, 2, 3], (), (), ["a", "b", "c"])
 
 
 @pytest.mark.parametrize(
     "key_cols, param_cols, val_cols",
-    [(["a"], ["b"], ["c"]), ([], ["b"], ["c", "a"]), ([], ["b"], ["a", "c"])],
+    [
+        (["a"], ["b"], ["c"]),
+        ([], ["b"], ["c", "a"]),
+        ([], ["b"], ["a", "c"]),
+        ([], ["b"], ["c"]),
+        (["a"], [], ["c"]),
+        (["a"], ["b"], []),
+        (["a"], [], []),
+        ([], ["b"], []),
+    ],
 )
 def test_validate_parameters_pass_dataframe(key_cols, param_cols, val_cols):
     data = pd.DataFrame({"a": [1, 2], "b_start": [0, 5], "b_end": [5, 10], "c": [100, 150]})
     validate_build_table_parameters(data, key_cols, param_cols, val_cols)
 
 
-class TestLookupTable(LookupTable):
-    def call(self, index: pd.Index) -> Union[pd.Series, pd.DataFrame]:
-        pass
-
-
 @pytest.mark.parametrize("validate", [True, False])
-def test_validate_option_invalid_data(validate, base_config):
-    simulation = InteractiveContext(components=[TestPopulation()], configuration=base_config)
-    manager = simulation._tables
+def test_validate_flag(mocker, validate):
+    manager = LookupTableManager()
+    manager.setup(mocker.Mock())
     manager._validate = validate
+    interface = LookupTableInterface(manager)
+
+    mock_validator = mocker.patch(
+        "vivarium.framework.lookup.manager.validate_build_table_parameters"
+    )
+
+    interface.build_table(0, value_columns=["a"])
+
     if validate:
-        with pytest.raises(ValueError, match="supply some data"):
-            manager._build_table([], [], [], [])
+        mock_validator.assert_called_once()
     else:
-        manager._build_table([], [], [], [])
-
-
-@pytest.mark.parametrize("validate", [True, False])
-def test_validate_option_valid_data(validate, base_config):
-    data = [1, 2, 3]
-    key_cols = ["KEY"]
-    param_cols = ["d"]
-    val_cols = ["a", "b", "c"]
-
-    simulation = InteractiveContext(components=[TestPopulation()], configuration=base_config)
-    manager = simulation._tables
-    manager._validate = validate
-
-    manager._build_table(data, key_cols, param_cols, val_cols)
+        mock_validator.assert_not_called()